Moved net/faraday assertions into their respective tests
